app/dash/requirements.txt
streamlit
pandas
Flask-SQLAlchemy
SQLAlchemy
pymysql
psycopg2-binary
plotly